% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.powRICLPM}
\alias{coef.powRICLPM}
\title{Extract results from powRICLPM analysis}
\usage{
\method{coef}{powRICLPM}(object, ..., parameter = NULL)
}
\arguments{
\item{object}{A powRICLPM object.}

\item{...}{Additional arguments not affecting the results extracted produced.}

\item{parameter}{A character string denoting the parameter of interest.}
}
\value{
A \code{data.frame} with rows representing the experimental conditions, and columns representing the metrics.
}
\description{
S3 method for class `powRICLPM`. \code{coef.powRICLPM} extracts the power analysis results across all experimental conditions, \emph{for a specific parameter}. See "Details" for a list of measures that are extracted.
}
\details{
The below measures are extracted for each experimental condition:
\itemize{
  \item \code{errors}: The number replications that ended with a fatal error.
  \item \code{not_converged}: The number of replications that did not converge to a solution.
  \item \code{inadmissible}: The number of replications that resulted in inadmissible results for some parameters (e.g., negative variances).
  \item \code{avg}: The average parameter estimate over all replications.
  \item \code{stdDev}: The standard deviation of parameter estimates across replications.
  \item \code{SEAvg}: The average standard error across replications.
  \item \code{mse}: The mean square error.
  \item \code{acc}: The average width of the confidence interval.
  \item \code{cover}: The coverage rate of the confidence interval for the defined significance criterion (by default \code{alpha = .05}).
  \item \code{pwr}: The proportion of times the \emph{p}-value was lower than the significance criterion.
}
}
