% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{coef_powRICLPM}
\alias{coef_powRICLPM}
\title{Performance Measures From `powRICLPM` Object}
\usage{
coef_powRICLPM(object, parameter)
}
\arguments{
\item{object}{A `powRICLPM` object.}

\item{parameter}{A character string denoting a single variable of interest. Use \code{\link{names_powRICLPM}} to get an overview of parameter names in the `powRICLPM` object.}
}
\value{
A `data.frame` object with columns containing performance measures, and rows representing experimental conditions.
}
\description{
\code{coef_powRICLPM} extracts performance measures (e.g., bias, mean square error, power) for a specific parameter, across all experimental conditions, from a `powRICLPM` object.
}
