% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-values.R
\name{compute_Psi}
\alias{compute_Psi}
\title{Compute Residual Variances of Lagged Within-Components}
\usage{
compute_Psi(Phi, wSigma)
}
\arguments{
\item{Phi}{A matrix of standardized autoregressive and cross-lagged effects in the population. Columns represent predictors and rows represent outcomes.}

\item{wSigma}{A correlation matrix for the within-unit components.}
}
\value{
A variance-covariance matrix for within-unit residuals from wave 2 and later.
}
\description{
Within an RI-CLPM context, this function computes the variance-covariance matrix of the within-unit residuals from wave 2 and later, given the lagged effects in \code{Phi} and an observed variance-covariance matrix \code{Sigma}.
}
\details{
The function is based on Equation (3.26) in Kim and Nelson (1999, p. 27).
}
