% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan.R
\name{create_lavaan}
\alias{create_lavaan}
\title{Create Lavaan Model Syntax for RICLPM Power Analysis}
\usage{
create_lavaan(
  time_points,
  RI_var = NULL,
  RI_cov = NULL,
  Phi = NULL,
  wSigma = NULL,
  Psi = NULL,
  syntax = FALSE,
  estimation = FALSE,
  constraints
)
}
\arguments{
\item{time_points}{An integer (vector) with elements larger than 3, indicating number of time points.}

\item{RI_var}{Numeric value denoting the random intercept variance.}

\item{RI_cov}{Numeric value denoting the covariance between the random intercepts.}

\item{Phi}{A matrix of standardized autoregressive and cross-lagged effects in the population. Columns represent predictors and rows represent outcomes.}

\item{wSigma}{A correlation matrix for the within-unit components.}

\item{Psi}{Variance-covariance matrix of within-unit residuals from wave 2 onwards.}

\item{syntax}{Logical indicating whether model syntax should be created.}

\item{estimation}{Logical indicator whether lavaan model syntax is inteded to simulate data. If FALSE, the lavaan syntax is created for the estimation of data.}

\item{constraints}{A character string specifying the RI-CLPM parameters that should be constrained over time in the estimation model. "\code{lagged}" refers to both cross-lagged and autoregressive effects, "\code{residuals}" refers to the residual variances, "\code{within}" refers to both lagged effects and residual variances, and "\code{stationarity}" imposes stationarity constraints on the lagged effects and residual variances. By default, the lagged effects and residual variances are allowed to vary freely over time.}
}
\value{
A data frame containing the model parameters (parameter elements as characters).
}
\description{
Creates a \pkg{lavaan} parameter table or model syntax for the specified model.
}
\details{
\subsection{Data generation}{If lavaan model syntax needs to be created for data generation, the user must provide values for the \code{ICC}, \code{RI_cor}, \code{Phi}, \code{wSigma}, and \code{Psi} arguments. By default, these arguments are set to \code{NULL}, such that when the model syntax is made from the parameter table, these parameters are estimated rather than set.}

\subsection{Naming conventions}{Details on the naming conventions can be found in the "Details" section of \code{\link{powRICLPM}}.}
}
