% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mplus.R
\name{create_Mplus}
\alias{create_Mplus}
\title{Create Mplus Model Syntax for RICLPM Power Analysis}
\usage{
create_Mplus(input)
}
\arguments{
\item{input}{A list with elements representing input for the Monte Carlo power analysis. See "Details" for the required elements in \code{input}.}
}
\description{
Creates Mplus model syntax for a Monte Carlo power analysis for the random intercept cross-lagged panel model (RI-CLPM).
}
\details{
\subsection{\code{Input}}{The \code{input} argument must contain the following elements:
\itemize{
  \item \code{sample_size}: The sample size.
  \item \code{time_points}: The number of time points.
  \item \code{ICC}: The proportion of between-unit variance.
  \item \code{RI_var}: The random intercept variance.
  \item \code{RI_cov}: The covariance between the random intercepts.
  \item \code{Phi}: A matrix of standardized autoregressive and cross-lagged effects.
  \item \code{wSigma}: A correlation matrix for the within-components.
  \item \code{Psi}: Residual variance-covariance matrix for the within-components.
  \item \code{reps}: Number of replications.
  \item \code{seed}: An integer of length 1.
  \item \code{save_path}: The directory (data) to save Mplus model syntax to.
  \item \code{constraints}: The constraints imposed on the estimation model.}}

\subsection{Naming conventions}{Details on the naming conventions can be found in the "Details" section of \code{\link{powRICLPM}}.}
}
