% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResVar.R
\name{getResVar}
\alias{getResVar}
\title{Compute Residual Variances of Lagged Within-Components}
\usage{
getResVar(effects, obsVar)
}
\arguments{
\item{effects}{Square matrix of standardized lagged effects of within-components}

\item{obsVar}{The observed variance-covariance matrix}
}
\value{
Residual variance of lagged within-components
}
\description{
This function only works for a stationary process.
}
\examples{
phi <- matrix(c(.45 , .1, .15, .15 ), ncol = 2, nrow = 2, byrow = T) # Lagged effects
Sigma <- Sig <- matrix(c(1 , .4, .4, 1 ), ncol = 2, nrow = 2, byrow = T) # Observed variance-covariance matrix
getResVar(phi, Sigma)
}
