% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lavaan.R
\name{lav_lagged}
\alias{lav_lagged}
\title{Create Lavaan Syntax for Within Lagged Effects}
\usage{
lav_lagged(input)
}
\arguments{
\item{input}{A list with objects necessary for creating lavaan model syntax. See Details for complete list of necessary elements.}
}
\value{
A data frame (parameter table) with lavaan model syntax for the within-unit lagged effects
}
\description{
Generate parameter table for the within-unit lagged effects of the RICLPM.
}
