% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lavaan.R
\name{lav_ME}
\alias{lav_ME}
\title{Create Lavaan Syntax No Measurement Errors}
\usage{
lav_ME(input)
}
\arguments{
\item{input}{A list with objects necessary for creating lavaan model syntax. See Details for complete list of necessary elements.}
}
\value{
A data frame (parameter table) with lavaan model syntax for unique factor variances (e.g., measurement errors) constrained to 0.
}
\description{
Create Lavaan Syntax No Measurement Errors
}
