% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lavaan.R
\name{lav_RI}
\alias{lav_RI}
\title{Create Lavaan Syntax for Random Intercepts}
\usage{
lav_RI(input)
}
\arguments{
\item{input}{A list with objects necessary for creating lavaan model syntax. See Details for complete list of necessary elements.}
}
\value{
A data frame (parameter table) with lavaan model syntax for the random intercepts.
}
\description{
Generate parameter table for the random intercepts of the RICLPM.
}
\details{
The \code{input} argument must containg the following elements:
\itemize{
  \item{$k}{Number of variables.}
  \item{$time_points}{The number of time points.}
  \item{$ICC}{Proportion of variance at the between-unit level.}
  \item{$RI_cor}{The random intercept correlations.}
  \item{$Phi}{Positive definite matrix of standardized autoregressive and cross-lagged effects in the population. Columns represent predictors and rows represent outcomes.}
  \item{wSigma}{Variance-covariance matrix of within-unit components. Variances of within-unit components must be set to 1 such that this matrix is equivalent to the correlation matrix.}
  \item{Psi}{Variance-covariance matrix of within-unit residuals at wave 2 and later.}
  \item{$name_obs}{Matrix with names of observed variables.}
  \item{$name_RI}{Matrix with names of random intercept factors.}
  \item{$name_within}{Matrix with names of within-components.}
}
}
