% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lavaan.R
\name{lav_RI_cor}
\alias{lav_RI_cor}
\title{Create Lavaan Syntax for Random Intercepts Covariance}
\usage{
lav_RI_cor(input)
}
\arguments{
\item{input}{A list with objects necessary for creating lavaan model syntax. See Details for complete list of necessary elements.}
}
\value{
A data frame (parameter table) with lavaan model syntax for the covariance(s) between the random intercepts.
}
\description{
Generate parameter table for the covariance(s) between the random intercepts of the RICLPM.
}
