% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lavaan.R
\name{lav_RI_var}
\alias{lav_RI_var}
\title{Create Lavaan Syntax for Random Intercepts Variance}
\usage{
lav_RI_var(input)
}
\arguments{
\item{input}{A list with objects necessary for creating lavaan model syntax. See Details for complete list of necessary elements.}
}
\value{
A data frame (parameter table) with lavaan model syntax for the variances of the random intercepts.
}
\description{
Generate parameter table for the variance of the random intercepts of the RICLPM.
}
