% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lavaan.R
\name{lav_within}
\alias{lav_within}
\title{Create Lavaan Syntax for Within Components}
\usage{
lav_within(input)
}
\arguments{
\item{input}{A list with objects necessary for creating lavaan model syntax. See Details for complete list of necessary elements.}
}
\value{
A data frame (parameter tabel) with lavaan model syntax for the within-unit components.
}
\description{
Generate parameter table for creating the within-unit components of the RICLPM.
}
