% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lavaan.R
\name{lav_within_cov1}
\alias{lav_within_cov1}
\title{Create Lavaan Syntax for Wave 1 Residual Covariance}
\usage{
lav_within_cov1(input)
}
\arguments{
\item{input}{A list with objects necessary for creating lavaan model syntax. See Details for complete list of necessary elements.}
}
\value{
A data frame (parameter table) with lavaa model syntax for the residual covariance(s) between the within-unit components of the RICLPM.
}
\description{
Generate parameter table for the residual covariance(s) between the within-unit components of the RICLPM.
}
