% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lavaan.R
\name{lav_within_cov2}
\alias{lav_within_cov2}
\title{Create Lavaan Syntax for Later Residual Covariance}
\usage{
lav_within_cov2(input)
}
\arguments{
\item{input}{A list with objects necessary for creating lavaan model syntax. See Details for complete list of necessary elements.}
}
\value{
A data frame (parameter table) with lavaan model syntax for the residual covariances of the RICLPM for wave 2 and later.
}
\description{
Create Lavaan Syntax for Later Residual Covariance
}
