% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lavaan.R
\name{lav_within_var1}
\alias{lav_within_var1}
\title{Create Lavaan Syntax for Wave 1 Residual Variance}
\usage{
lav_within_var1(input)
}
\arguments{
\item{input}{A list with objects necessary for creating lavaan model syntax. See Details for complete list of necessary elements.}
}
\value{
A data frame (parameter table) with lavaan model syntax for residual variances of the within-unit components at wave 1.
}
\description{
Generate parameter table for the residual variance of the within-unit components at wave 1 of the RICLPM.
}
