% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_Mplus.R
\name{Mplus_lagged}
\alias{Mplus_lagged}
\title{Create Mplus Syntax for Within Lagged Effects}
\usage{
Mplus_lagged(input, estimation = F)
}
\arguments{
\item{input}{A list with objects necessary for creating Mplus model syntax. See Details for complete list of necessary elements.}
}
\value{
A data frame (parameter table) with Mplus syntax for the within-unit lagged effects.
}
\description{
Create Mplus Syntax for Within Lagged Effects
}
