% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_Mplus.R
\name{Mplus_ME}
\alias{Mplus_ME}
\title{Create Mplus Syntax No Measurement Errors}
\usage{
Mplus_ME(input)
}
\arguments{
\item{input}{A list with objects necessary for creating Mplus model syntax. See Details for complete list of necessary elements.}
}
\value{
A data frame (parameter table) with Mplus syntax for constraining the unique factor variances of the observed variables (e.g., measurement error variances) to 0.
}
\description{
Create Mplus Syntax No Measurement Errors
}
