% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_Mplus.R
\name{Mplus_RI}
\alias{Mplus_RI}
\title{Create Mplus Syntax for Random Intercepts}
\usage{
Mplus_RI(input)
}
\arguments{
\item{input}{A list with objects necessary for creating Mplus model syntax. See Details for complete list of necessary elements.}
}
\value{
A data frame (parameter table) with Mplus syntax for random intercepts.
}
\description{
Create Mplus Syntax for Random Intercepts
}
