% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_Mplus.R
\name{Mplus_RI_cor}
\alias{Mplus_RI_cor}
\title{Create Mplus Syntax for Random Intercept Covariance}
\usage{
Mplus_RI_cor(input, estimation = F)
}
\arguments{
\item{input}{A list with objects necessary for creating Mplus model syntax. See Details for complete list of necessary elements.}
}
\value{
A data frame (parameter table) with Mplus syntax for the covariance(s) between the random intercepts.
}
\description{
Create Mplus Syntax for Random Intercept Covariance
}
