% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_Mplus.R
\name{Mplus_RI_var}
\alias{Mplus_RI_var}
\title{Create Mplus Syntax for Random Intercept Variance}
\usage{
Mplus_RI_var(input, estimation = F)
}
\arguments{
\item{input}{A list with objects necessary for creating Mplus model syntax. See Details for complete list of necessary elements.}
}
\value{
A data frame (parameter table) with Mplus syntax for the variances of the random intercepts.
}
\description{
Create Mplus Syntax for Random Intercept Variance
}
