% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_Mplus.R
\name{Mplus_within}
\alias{Mplus_within}
\title{Create Mplus Syntax for Within Components}
\usage{
Mplus_within(input)
}
\arguments{
\item{input}{A list with objects necessary for creating Mplus model syntax. See Details for complete list of necessary elements.}
}
\value{
A data frame (parameter table) with Mplus syntax for the within components.
}
\description{
Create Mplus Syntax for Within Components
}
