% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_Mplus.R
\name{Mplus_within_var1}
\alias{Mplus_within_var1}
\title{Create Mplus Syntax for Wave 1 Within-Component Variance}
\usage{
Mplus_within_var1(input, estimation = F)
}
\arguments{
\item{input}{A list with objects necessary for creating Mplus model syntax. See Details for complete list of necessary elements.}
}
\value{
A data frame (parameter table) with Mplus syntax for the within-unit variance at wave 1.
}
\description{
Create Mplus Syntax for Wave 1 Within-Component Variance
}
