% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_Mplus.R
\name{Mplus_within_var2}
\alias{Mplus_within_var2}
\title{Create Mplus Syntax for Later Residual Variance}
\usage{
Mplus_within_var2(input, estimation = F)
}
\arguments{
\item{input}{A list with objects necessary for creating Mplus model syntax. See Details for complete list of necessary elements.}
}
\value{
A data frame (parameter table) with Mplus syntax for within-unit residual variances at wave 2 and later.
}
\description{
Create Mplus Syntax for Later Residual Variance
}
