% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{names_powRICLPM}
\alias{names_powRICLPM}
\title{Parameter Names From `powRICLPM` Object}
\usage{
names_powRICLPM(object)
}
\arguments{
\item{object}{A `powRICLPM` object.}
}
\value{
A character vector with the names of the variables internally created by the \pkg{powRICLPM} package.
}
\description{
\code{names_powRICLPM} extracts the names of the variables that are internally created by the powRICLPM package. Details about the naming conventions can be found in the "Details" section of \code{\link{powRICLPM}}.
}
\details{
When simulating the power for conditions with a varying number of time points, there are different amounts of parameters across the conditions. By default, this function returns the parameter names from the condition with the smallest number of parameters, such that the returned parameter names are valid for each condition.
}
\examples{
# Define population parameters for lagged effects and within-component correlations
Phi <- matrix(c(.4, .1, .2, .3), ncol = 2, byrow = TRUE)
wSigma <- matrix(c(1, .3, .3, 1), ncol = 2, byrow = TRUE)

# Create powRICLPM object for à priori power analysis
output <- powRICLPM(target_power = 0.5,
                    search_lower = 300,
                    search_upper = 600,
                    search_step = 50,
                    time_points = 3,
                    ICC = 0.5,
                    RI_cor = 0.3,
                    Phi = Phi,
                    wSigma = wSigma,
                    reps = 50,
                    seed = 123456)

# Get names of internally created parameters
names_powRICLPM(output)
}
