% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.powRICLPM}
\alias{plot.powRICLPM}
\title{Plot `powRICLPM` analysis results}
\usage{
\method{plot}{powRICLPM}(x, y, ..., parameter)
}
\arguments{
\item{x}{A `powRICLPM` object.}

\item{y}{Argument not in use.}

\item{...}{Argument not in use.}

\item{parameter}{A character string denoting the parameter of interest. Use the argument \code{names = TRUE} in \code{summary()} to get an overview of parameter names in the `powRICLPM` object.}
}
\value{
A `ggplot2` object.
}
\description{
\code{plot.powRICLPM} visualizes results from `powRICLPM` objects using \pkg{ggplot2}. The plots display the relation between sample size (x-axis) and power (y-axis), grouped by the number of time points, and wrapped by the proportion of between-unit variance.
}
\details{
\subsection{Manually creating plots}{\code{plot.powRICLPM()} creates the most obvious plot in the context of power analysis: The relation between power and sample size. However, \code{powRICLPM} computes several other metrics (e.g., accuracy, coverage rate, etc.) that research might want to plot. To this end, I suggest that user use \code{coef.powRICLPM} to collect all metrics, across all experimental conditions, for a specific parameter, into a data frame first. Then, user can use any plotting function they like to create their own manual plot. See "Details" of \code{\link{coef.powRICLPM}} for a list of measures that are computed by the function.}
}
\seealso{
\itemize{
  \code{\link{coef.powRICLPM}}: Extract performance measures for a specific parameter, across all experimental conditions. This function is used internally in \code{plot.powRICLPM}.
}
}
