% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_powRICLPM}
\alias{plot_powRICLPM}
\title{Plot `powRICLPM` Objects}
\usage{
plot_powRICLPM(object, x, y, ..., wrap, parameter = NULL)
}
\arguments{
\item{object}{A `powRICLPM` object.}

\item{x}{A character string denoting the variable to be plotted on the x-axis: Must be "sample_size", "time_points", or "ICC".}

\item{y}{A character string denoting the performance measure to be plotted on the y-axis. See "Details" for a list of valid performance measures.}

\item{...}{Additional options that are parsed to the mapping `aes()`.}

\item{wrap}{A character string denoting the variable to facet the plots by.}

\item{parameter}{A character string denoting a single variable of interest. Use \code{\link{names_powRICLPM}} to get an overview of parameter names in the `powRICLPM` object.}
}
\value{
A `ggplot2` object.
}
\description{
\code{plot_powRICLPM} visualizes results from `powRICLPM` objects using \pkg{ggplot2}.
}
\details{
\subsection{Performance measures}{\code{powRICLPM()} computes several performance measures that can be plotted for each condition. These include:
\itemize{
  \item \code{avg}: Average parameter estimate.
  \item \code{stdDev}: Standard deviation of the estimates.
  \item \code{SEAvg}: Mean standard error.
  \item \code{mse}: Mean square error.
  \item \code{acc}: Accuracy, the width of the confidence interval.
  \item \code{cover}: The proportion of times the confidence interval captures the true population value.
  \item \code{pwr}: The proportion of time the **p**-value is below the significance criterion.
  }
}

\subsection{Recommendations}{It is recommended to only include study design characteristics on the x-axis (e.g., sample size and the number of repeated measures), and to facet wrap by characteristics of the data (e.g., ICC, skewness, kurtosis). This strategy emphasizes that facets represent characteristics that influence power, but cannot be influenced by the researcher. In other words, facet wraps represent different "worlds", whereas the factors on the x-axis can be tweaked by researchers to reach the desired level of power.}
}
\seealso{
\itemize{
  \code{\link{coef_powRICLPM}}: Extract performance measures for a specific parameter, across all experimental conditions. This function is used internally in \code{plot_powRICLPM}.
}
}
