% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powRICLPM-package.R
\docType{package}
\name{powRICLPM-package}
\alias{powRICLPM-package}
\title{powRICLPM: Perform Power Analysis for the RI-CLPM and STARTS Model}
\description{
Perform user-friendly power analyses for the random intercept cross-lagged panel model (RI-CLPM) and the bivariate stable trait autoregressive trait state (STARTS) model. The strategy as proposed by Mulder (2023) \doi{10.1080/10705511.2022.2122467} is implemented. Extensions include the use of parameter constraints over time, bounded estimation, generation of data with skewness and kurtosis, and the option to setup the power analysis for Mplus.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jeroendmulder.github.io/powRICLPM/}
  \item Report bugs at \url{https://github.com/JeroenDMulder/powRICLPM/issues/}
}

}
\author{
\strong{Maintainer}: Jeroen Mulder \email{j.d.mulder@uu.nl} (\href{https://orcid.org/0000-0002-5553-0856}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Netherlands Organization for Scientific Research [funder]
}

}
\keyword{internal}
