% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{powRICLPM_names}
\alias{powRICLPM_names}
\title{Get Parameter Names From "powriclpm-Object}
\usage{
powRICLPM_names(x, max_set = FALSE)
}
\arguments{
\item{x}{A powRICLPM object.}

\item{max_set}{A logical indicating if the parameter names from the condition with the largest number of parameters should be returned. When simulating the power for conditions with varying number of time points, then there are different numbers of parameters over the conditions. By default, this function returns the parameter names from the condition with the smallest number of parameters, such that the returned parameter names are valid for each condition.}
}
\value{
A character vector with the names of the variables internally created by the powRICLPM package.
}
\description{
\code{\link{powRICLPM_names}} gets the names of the variables that are internally created by the powRICLPM package. Details about the naming conventions can be found in the "Details" section of \code{\link{powRICLPM}}.
}
\examples{
# Define population parameters for lagged effects and within-component correlations
Phi <- matrix(c(.4, .1, .2, .3), ncol = 2, byrow = TRUE)
wSigma <- matrix(c(1, .3, .3, 1), ncol = 2, byrow = TRUE)

# Create powRICLPM object for à priori power analysis
output <- powRICLPM(target_power = 0.5,
                     parameter = "wB2~wA1",
                     search_lower = 300,
                     search_upper = 600,
                     search_step = 50,
                     time_points = 3,
                     ICC = 0.5,
                     RI_cor = 0.3,
                     Phi = Phi,
                     wSigma = wSigma,
                     reps = 50,
                     cores = 1,
                     seed = 123456)

# Get names of internally created parameters
powRICLPM_names(output)
}
\seealso{
\code{\link{powRICLPM_summary}}
}
