% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.powRICLPM}
\alias{print.powRICLPM}
\title{Print `powRICLPM` object}
\usage{
\method{print}{powRICLPM}(x, ...)
}
\arguments{
\item{x}{A `powRICLPM` object.}

\item{...}{Argument not in use.}
}
\description{
\code{print.powRICLPM} prints a textual summary of the power analysis \emph{design} used to create the `powRICLPM` object. Specifically, it prints the number of Monte Carlo replications that were run, the range of sample sizes, number of repeated measures, and proportions of between-unit variance, and if bounded estimation, constraints over time, or informative hypotheses were used.
}
