% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_condition}
\alias{run_condition}
\title{Run Monte Carlo Simulation For Single Condition}
\usage{
run_condition(object, p)
}
\arguments{
\item{object}{A list with information for running a single simulation condition. See "Details" for an overview of the elements this object must contain.}

\item{p}{A \pkg{progressr} object}
}
\value{
A list.
}
\description{
\code{run_condition()} runs a Monte Carlo simulation for a single given condition. It generates data based on the \code{pop_synt} element in "object", and estimates an RI-CLPM using the \code{est_synt} element in "object". Data generation and model estimation are done using \pkg{lavaan}.
}
\details{
\subsection{Input: Elements in "object"}{To successfully run a Monte Carlo simulation, the "object" arguments needs the following elements:
\itemize{
  \item \code{sample_size}: The sample size.
  \item \code{time_points}: The number of time points.
  \item \code{reps}: Number of replications.
  \item \code{pop_synt}: \pkg{lavaan} model syntax containing population values for data generation.
  \item \code{pop_tab}: \pkg{lavaan} parameter table for data generation.
  \item \code{est_synt}: \pkg{lavaan} model syntax for estimation.
  \item \code{est_tab}: \pkg{lavaan} parameter table for estimation.
  \item \code{skewness}: The skewness value(s) for the observed variables.
  \item \code{kurtosis}: The kurtosis value(s) for the observed variables.
  \item \code{alpha}: The significance criterion.
  \item \code{save_path}: Folder to which simulated data are saved (optional).
  }
}

\subsection{Output}{This function adds the following elements to "object":
\itemize{
  \item \code{results}: A data frame containing the results (i.e., population values, bias, standard error of the estimate, coverage, power, etc.),
  \item \code{errors}: A logical vector denoting failed Monte Carlo replications,
  \item \code{not_converged}: A logical vector denoting non-converged Monte Carlo replications, and
  \item \code{inadmissible} A logical vector denoting Monte Carlo replications that resulted in negative variances or non-positive definite matrices.
  }
}
}
