% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.powRICLPM}
\alias{summary.powRICLPM}
\title{Summarize setup and results from \code{powRICLPM} object}
\usage{
\method{summary}{powRICLPM}(
  object,
  ...,
  parameter = NULL,
  sample_size = NULL,
  time_points = NULL,
  ICC = NULL
)
}
\arguments{
\item{object}{A \code{powRICLPM} object.}

\item{...}{(don't use) Additional arguments not affecting the summary produced.}

\item{parameter}{Character string of length denoting the parameter to visualize the results for.}

\item{sample_size}{(optional) An integer (vector), indicating specific sample sizes at which to evaluate power, rather than specifying a range using the \code{search_} arguments.}

\item{time_points}{An integer (vector) with elements at least larger than 3, indicating number of time points.}

\item{ICC}{A \code{double} (vector), denoting the proportion of variance at the between-unit level.}
}
\description{
S3 method for class \code{powRICLPM}. \code{summary.powRICLPM} summarizes and outputs the setup and results of the \code{powRICLPM} analysis. Depending on the input, \code{summary.powRICLPM} provides a different summary (see "Details").
}
\details{
\code{summary.powRICLPM} provides a different summary of the \code{powRICLPM} object, depending on the additional arguments that are set:
\itemize{
  \item When \code{sample_size = ...}, \code{time_points = ...}, and \code{ICC = ...} are set: Estimation information and results for all parameters of the experimental condition denoted by \code{sample_size}, \code{time_points}, and \code{ICC}.
  \item When \code{parameter = "..."} is set: Estimation information and results for a specific parameter across all experimental conditions.
  \item No additional arguments: Characteristics of the different experimental conditions are summarized, as well as session info (information that applies to each conditions, such the number of replications, etc.).
}
}
\examples{
# Example - Same starting point as example from ?powRICLPM()
# Define population parameters for lagged effects and within-component correlations
Phi <- matrix(c(.4, .1, .2, .3), ncol = 2, byrow = TRUE)
wSigma <- matrix(c(1, .3, .3, 1), ncol = 2, byrow = TRUE)

# Setup parallel computing (multicore, speeding up the analysis)
\dontrun{
library(furrr)
plan(multisession, workers = 4)

# Run the analysis
out1 <- powRICLPM(
  target_power = 0.8,
  search_lower = 500,
  search_upper = 1000,
  search_step = 100,
  time_points = c(3, 4),
  ICC = c(0.4, 0.5, 0.6),
  RI_cor = 0.3,
  Phi = Phi,
  wSigma = wSigma,
  reps = 100,
  seed = 123456
)

# General results (i.e., setup of powRICLPM analysis)
summary(out)

# General results and recommendation for specific parameter
summary(out, parameter = "wB2~wA1")

# Summary for specific condition
summary(out, sample_size = 600, time_points = 4, ICC = .5)
}
}
