% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lavaan.R
\name{create_lavaan}
\alias{create_lavaan}
\title{Create Lavaan Model Syntax for RICLPM Power Analysis}
\usage{
create_lavaan(
  time_points,
  ICC = NULL,
  RI_cor = NULL,
  Phi = NULL,
  wSigma = NULL,
  Psi = NULL,
  syntax = F
)
}
\arguments{
\item{time_points}{An integer or integer vector with elements larger than 3, indicating number of time points.}

\item{ICC}{A numeric value denoting the proportion of variance at the between-unit level.}

\item{RI_cor}{A numeric value denoting the correlation between random intercepts.}

\item{Phi}{A matrix of standardized autoregressive and cross-lagged effects in the population. Columns represent predictors and rows represent outcomes.}

\item{wSigma}{A correlation matrix for the within-unit components.}

\item{Psi}{Variance-covariance matrix of within-unit residuals from wave 2 onwards.}

\item{syntax}{Logical indicating whether model syntax should be created.}
}
\value{
A data frame containing the model parameters (parameter elements as characters).
}
\description{
Creates a \pkg{lavaan} parameter table or model syntax for the specified model.
}
\details{
\subsection{Data generation}
If lavaan model syntax needs to be created for data generation, the user must provide values for the \code{ICC}, \code{RI_cor}, \code{Phi}, \code{wSigma}, and \code{Psi} arguments. By default, these arguments are set to \code{NULL}, such that when the model syntax is made from the parameter table, these parameters are estimated rather than set.

\subsection{Naming conventions}
Details on the naming conventions can be found in the "Details" section of \code{\link{powRICLPM}}.
}
\examples{
# Define population parameters for lagged effects and within-component correlations
Phi <- matrix(c(.5, .1, .4, .5), ncol = 2, byrow = T)
wSigma <- matrix(c(1 , .3, .3, 1) , ncol = 2, byrow = T)
Psi <- compute_Psi(Phi = Phi, wSigma = wSigma)

# Create lavaan model syntax for data generation
syntax <- create_lavaan(time_points = 3,
                        ICC = 0.5,
                        RI_cor = 0.3,
                        Phi = Phi,
                        wSigma = wSigma,
                        Psi = Psi,
                        syntax = F)
}
