% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.powRICLPM}
\alias{plot.powRICLPM}
\title{Plot powRICLPM Objects}
\usage{
\method{plot}{powRICLPM}(x, y = NULL, parameter = NULL)
}
\arguments{
\item{x}{A powRICLPM object.}

\item{y}{Phantom argument that has no influence on the plot produced. Included here to make the generic \code{plot()} a method for objects of class "powRICLPM".}

\item{parameter}{A character string denoting a single variable of interest. Use \code{\link{names.powRICLPM}} to get an overview of valid parameter names for the powRICLPM object.}
}
\description{
Plotting method for powRICLPM objects. This function visualizes results from à priori and post hoc powRICLPM analyses.
}
\examples{
# Define population parameters for lagged effects and within-component correlations
Phi <- matrix(c(.4, .1, .2, .3), ncol = 2, byrow = T)
wSigma <- matrix(c(1, .3, .3, 1), ncol = 2, byrow = T)

# Create powRICLPM object for à priori power analysis
output <- powRICLPM(target_power = 0.5,
                     parameter = "wB2~wA1",
                     search_lower = 300,
                     search_upper = 600,
                     search_step = 50,
                     time_points = 3,
                     ICC = 0.5,
                     RI_cor = 0.3,
                     Phi = Phi,
                     wSigma = wSigma,
                     reps = 50,
                     cores = 1,
                     seed = 123456)

# Visualize results
plot(output, parameter = "wB2~wA1")

}
