% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powRICLPM.R
\name{powRICLPM}
\alias{powRICLPM}
\title{Power analysis for the RI-CLPM}
\usage{
powRICLPM(
  target_power,
  search_lower = NULL,
  search_upper = NULL,
  search_step = 20,
  sample_size = NULL,
  time_points,
  ICC,
  RI_cor,
  Phi,
  wSigma,
  reliability = 1,
  skewness = 0,
  kurtosis = 0,
  est_ME = FALSE,
  alpha = 0.05,
  reps = 20,
  bootstrap_reps = 1000,
  seed = NA,
  constraints = "none",
  bounds = FALSE,
  estimator = NA
)
}
\arguments{
\item{target_power}{A numeric value between 0 and 1, denoting the targeted power level.}

\item{search_lower}{A positive integer, denoting the lower bound of a range of sample sizes.}

\item{search_upper}{A positive integer, denoting the upper bound of a range of sample sizes.}

\item{search_step}{A positive integer, denoting an increment in sample size.}

\item{sample_size}{(optional) An integer (vector), indicating specific sample sizes at which to evaluate power, rather than specifying a range using the \code{search_} arguments.}

\item{time_points}{An integer (vector) with elements at least larger than 3, indicating number of time points.}

\item{ICC}{A \code{double} (vector), denoting the proportion of variance at the between-unit level.}

\item{RI_cor}{A \code{double}, denoting the correlation between random intercepts.}

\item{Phi}{A matrix, with standardized autoregressive effects (on the diagonal) and cross-lagged effects (off-diagonal) in the population. Columns represent predictors and rows represent outcomes.}

\item{wSigma}{A correlation matrix for the within-unit components.}

\item{reliability}{(optional) A numeric value between 0 and 1, denoting the reliability of the variables.}

\item{skewness}{(optional) A numeric value, denoting the skewness values for the observed variables (see \code{\link[lavaan]{simulateData}}).}

\item{kurtosis}{(optional) A numeric value, denoting the excess kurtosis values (i.e., compared to the kurtosis of a normal distribution) for the observed variables (see \code{\link[lavaan]{simulateData}}).}

\item{est_ME}{(optional) A logical, denoting if measurement error variance should be estimated in the RI-CLPM (see "Details").}

\item{alpha}{(optional) A \code{double}, denoting the significance criterion.}

\item{reps}{A positive \code{integer}, denoting the number of Monte Carlo replications to be used during simulations.}

\item{bootstrap_reps}{(optional) A positive \code{integer}, denoting the number of bootstrap samples to use for quantifying the uncertainty around the power analysis results.}

\item{seed}{An `integer` of length 1. If multiple cores are used, a seed of length 1 will be used to generate a full L'Ecuyer-CMRG seed for all cores (see \code{\link[furrr]{furrr_options}}).}

\item{constraints}{(optional) A character string, specifying the type of constraints that should be imposed on the estimation model (see "Details").}

\item{bounds}{(optional) A logical, denoting if bounded estimation should be used for the latent variable variances in the model (see "Details").}

\item{estimator}{(options) A character, denoting the estimator to be used (see "Details").}
}
\value{
A list containing a \code{conditions} and \code{session} element. \code{condition} itself is a list of experimental conditions, where each element is again a list containing the input and output of the power analysis for that particular experimental condition. \code{session} is a list containing information common to all experimental conditions.
}
\description{
\code{powRICLPM()} performs a Monte Carlo power analysis for the random intercept cross-lagged panel model (RI-CLPM). It computes performance metrics (e.g., bias, mean square error, coverage, power, etc) for all RI-CLPM parameters, and can perform power analysis across multiple experimental conditions simultaneously. Conditions are defined in terms of sample size, number of time points, and proportion of between-unit variance (ICC). See "Details" for information on a) the data simulation and model estimation, b) internal naming conventions of parameters, c) the option to include measurement errors and various constraints, and d) parallel execution capabilities for speeding up the analysis.
}
\details{
A rationale for the power analysis strategy implemented in this package can be found in Mulder (2022).

\subsection{Data generation and model estimation}{Data are generated using \code{\link[lavaan]{simulateData}} from the \pkg{lavaan} package. Based on \code{Phi} and \code{wSigma}, the residual variances and covariances for the within-components at wave 2 and later are computed, such that the within-components themselves have a variance of 1. This implies that the lagged effects in \code{Phi} can be interpreted as standardized effects.

Data are analyzed using \code{\link[lavaan]{lavaan}} from the \pkg{lavaan} package. The default estimator is maximum likelihood. Other maximum likelihood based estimators implemented in \href{https://lavaan.ugent.be/tutorial/est.html}{\pkg{lavaan}} can be specified as well. When skewed or kurtosed data are generated (using the \code{skewness} and \code{kurtosis} arguments), the estimator defaults to robust maximum likelihood (i.e., \code{estimator = "MLR"}). The population parameter values are used as starting values. Parameter estimates from unconverged model solutions are discarded from the results. However, parameter estimates from solutions with inadmissible parameter estimates (e.g., a negative random intercept variance), are included in the results. The results include the minimum estimates for all parameters across replications. This can be used to diagnose which parameter is the cause of the inadmissible solution.}

\subsection{Naming conventions for observed and latent variables}{The observed variables in the RI-CLPM are given default names, namely capital letters in alphabetical order, with numbers denoting the measurement occasion. For example, for a bivariate RICLPM with 3 time points, we observe \code{A1}, \code{A2}, \code{A3}, \code{B1}, \code{B2}, and \code{B3}. Their within-components are denoted by \code{wA1}, \code{wA2}, ..., \code{wB3}, respectively. The between-components have \code{RI_} prepended to the variable name, resulting in \code{RI_A} and \code{RI_B}.

Parameters are denoted using \pkg{lavaan} model syntax (see \href{https://lavaan.ugent.be/tutorial/syntax1.html}{the \pkg{lavaan} website}). For example, the random intercept variances are denoted by \code{RI_A~~RI_A} and \code{RI_B~~RI_B}, the cross-lagged effects at the first wave as \code{wB2~wA1} and \code{wA2~wB1}, and the autoregressive effects as \code{wA2~wA1} and \code{wB2~wB1}. Use \code{give(object, "names")} to extract parameter names from the \code{powRICLPM} object.}

\subsection{Imposing constraints}{The following constraints can be imposed on the estimation model using the \code{constraints = "..."} argument:

\itemize{
  \code{lagged}: Time-invariant autoregressive and cross-lagged effects.
  \code{residuals}: Time-invariant residual variances.
  \code{within}: Time-invariant lagged effects and residual variances.
  \code{stationarity}: Constraints such that at the within-unit level a stationary process is estimated. This included time-invariant lagged effects, and constraints on the residual variances.
  \code{ME}: Time-invariant measurement error variances. Only possible when \code{est_ME = TRUE}.
}
}

\subsection{Bounded estimation}{Bounded estimation is useful to avoid nonconvergence in small samples (< 100). Here, automatic wide bounds are used as advised by \href{https://doi.org/10.1080/10705511.2021.1982716}{De Jonckere and Rosseel (2022}; see \code{optim.bounds} in \code{\link[lavaan]{lavOptions}}). This option can only be used when no constraints are imposed on the estimation model.}

\subsection{Parallel processing using \pkg{furrr}}{To speed up the analysis, power analysis for multiple experimental conditions can be executed in parallel. This has been implemented using \pkg{furrr}. By default the analysis is executed sequentially (i.e., single-core). Parallel execution (i.e., multicore) can be setup using \code{\link[future]{plan}}, for example \code{plan(multisession, workers = 4)}. For more information and options, see \url{https://furrr.futureverse.org}.}

\subsection{Progress bar using \pkg{progressr}}{A progress bar displaying the status of the power analysis has been implemented using \pkg{progressr}. By default, a simple progress bar will be shown. For more information on how to control this progress bar and several other notification options (e.g., auditory notifications), see \url{https://progressr.futureverse.org}.}
}
\examples{
# Example - Simulate power across range of sample sizes
# Define population parameters for lagged effects and within-component correlations
Phi <- matrix(c(.4, .1, .2, .3), ncol = 2, byrow = TRUE)
wSigma <- matrix(c(1, .3, .3, 1), ncol = 2, byrow = TRUE)

# Setup parallel computing (multicore, speeding up the analysis)
\dontrun{
library(furrr)
library(progressr)
plan(multisession, workers = 4)

# Run preliminary analysis (with limited "reps")
with_progress({
  out1 <- powRICLPM(
    target_power = 0.8,
    search_lower = 500,
    search_upper = 1000,
    search_step = 100,
    time_points = c(3, 4),
    ICC = c(0.4, 0.5, 0.6),
    RI_cor = 0.3,
    Phi = Phi,
    wSigma = wSigma,
    reps = 50,
    seed = 1234
  )
})
}
}
\seealso{
\itemize{
  \item \code{\link{powRICLPM_Mplus}}: Create Mplus model syntax for RI-CLPM power analysis.
  \item \code{\link{summary}}: Summarize the setup of \code{powRICLPM} object.
  \item \code{\link{give}}: Extract information from \code{powRICLPM} objects.
  \item \code{\link{plot}}: Visualize results \code{powRICLPM} object for a specific parameter.
}
}
\author{
Jeroen D. Mulder \email{j.d.mulder@uu.nl}
}
