% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powRICLPM.R
\name{powRICLPM_Mplus}
\alias{powRICLPM_Mplus}
\title{(superseded) Create Mplus Syntax for RI-CLPM Power Analysis}
\usage{
powRICLPM_Mplus(...)
}
\arguments{
\item{...}{(don't use)}
}
\description{
The function \code{powRICLPM_Mplus} has been superseded and is no longer available in the current version of the package.
}
\details{
The functionality previously provided by \code{powRICLPM_Mplus} has been replaced by \code{powRICLPM}, where you can now set \code{software = "Mplus"}.
}
\examples{
\dontrun{
  # Use `software = "Mplus"` to setup power analysis for Mplus
  out_preliminary <- powRICLPM(
    target_power = 0.8,
    search_lower = 500,
    search_upper = 700,
    search_step = 100,
    time_points = c(3, 4),
    ICC = c(0.4, 0.6),
    reliability = c(1, 0.8),
    RI_cor = 0.3,
    Phi = Phi,
    within_cor = 0.3,
    reps = 1000,
    seed = 1234,
    software = "Mplus"
  )
}
}
\seealso{
\code{\link[=powRICLPM]{powRICLPM()}}
}
\keyword{internal}
