% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_Mplus.R
\name{powRICLPM_Mplus}
\alias{powRICLPM_Mplus}
\title{Create Mplus Syntax for RICLPM Power Analysis}
\usage{
powRICLPM_Mplus(
  sample_size,
  time_points,
  ICC,
  RI_cor,
  Phi,
  wSigma,
  Psi = NULL,
  reps = 1000,
  seed = NULL,
  save_path
)
}
\arguments{
\item{sample_size}{An integer or integer vector with elements larger than 1, indicating the sample sizes at which to evaluate power. Use only for post hoc analysis.}

\item{time_points}{An integer or integer vector with elements larger than 3, indicating number of time points.}

\item{ICC}{A numeric value denoting the proportion of variance at the between-unit level.}

\item{RI_cor}{A numeric value denoting the correlation between random intercepts.}

\item{Phi}{A matrix of standardized autoregressive and cross-lagged effects in the population. Columns represent predictors and rows represent outcomes.}

\item{wSigma}{A correlation matrix for the within-unit components.}

\item{Psi}{Variance-covariance matrix of within-unit residuals from wave 2 onwards.}

\item{reps}{A positive integer denoting the number of Monte Carlo replications to be used during simulations.}

\item{seed}{An integer of length 1. If multiple cores are used, a seed of length 1 will be used to generate a full L'Ecuyer-CMRG seed for all cores. For more information, see \code{\link[furrr]{furrr_options}}.}

\item{save_path}{A character string naming the directory to save any (data) files to.}
}
\value{
A character containing the Mplus model syntax.
}
\description{
\code{powRICLPM_Mplus()} is used to create and save \code{Mplus} model syntax for the specified RI-CLPM model for performing a Monte Carlo power analysis in Mplus.
}
\details{
\subsection{Syntax generation}
Mplus model syntax is created in two steps: First, the \code{MODEL POPULATION} command syntax is created in which parameters are constrained to population values, and second the \code{MODEL} command syntax is created for model estimation. Ultimately, the parameter tables are pasted together to form character vectors containing the Mplus syntax to be exported.

\subsection{Naming conventions}
Details on the naming conventions can be found in the "Details" section of \code{\link{powRICLPM}}.
}
\examples{
# Define population parameters for lagged effects and within-component correlations
Phi <- matrix(c(.5, .1, .4, .5), ncol = 2, byrow = TRUE)
wSigma <- matrix(c(1 , .3, .3, 1) , ncol = 2, byrow = TRUE)
Psi <- compute_Psi(Phi = Phi, wSigma = wSigma)

# Create and save Mplus model syntax for Monte Carlo power analysis
powRICLPM_Mplus(sample_size = 300,
                time_points = 3,
                ICC = 0.5,
                RI_cor = 0.3,
                Phi = Phi,
                wSigma = wSigma,
                reps = 10000,
                seed = 123456,
                save_path = "./saved")
}
