% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.powRICLPM}
\alias{print.powRICLPM}
\title{Print \code{powRICLPM} object}
\usage{
\method{print}{powRICLPM}(x, ...)
}
\arguments{
\item{x}{A \code{powRICLPM} object.}

\item{...}{Argument not in use.}
}
\description{
\code{print.powRICLPM} prints a table listing all experimental conditions contained in the \code{powRICLPM} object, as well as the frequency of the estimation problems that occurred in each.
}
