% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.powRICLPM}
\alias{print.powRICLPM}
\title{Print \code{powRICLPM} object}
\usage{
\method{print}{powRICLPM}(x, ...)
}
\arguments{
\item{x}{A \code{powRICLPM} object.}

\item{...}{Argument not in use.}
}
\description{
\code{print.powRICLPM} prints a textual summary of the setup of the power analysis within the \code{powRICLPM} object. Specifically, it prints the characteristics of the different experimental conditions, as well as session info (information that applies to every condition).
}
