% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.powRICLPM}
\alias{print.powRICLPM}
\title{Print \code{powRICLPM} Object}
\usage{
\method{print}{powRICLPM}(x, ...)
}
\arguments{
\item{x}{A \code{powRICLPM} object.}

\item{...}{(don't use)}
}
\value{
No return value, called for side effects.
}
\description{
\code{print.powRICLPM} prints a table listing all experimental conditions contained in the \code{powRICLPM} object, as well as the frequency of the estimation problems that occurred in each.
}
