% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.powRICLPM}
\alias{summary.powRICLPM}
\title{Summarize powRICLPM analysis}
\usage{
\method{summary}{powRICLPM}(
  object,
  ...,
  names = FALSE,
  parameter = NULL,
  sample_size = NULL,
  time_points = NULL,
  ICC = NULL
)
}
\arguments{
\item{object}{A powRICLPM object.}

\item{...}{Additional arguments not affecting the summary produced.}

\item{names}{A logical denoting if parameter names contained in \code{object} should be printed.}

\item{parameter}{A character string denoting the parameter of interest.}

\item{sample_size}{An integer denoting the sample size of interest.}

\item{time_points}{An integer denoting the number of time points of interest.}

\item{ICC}{A numeric value denoting the proportion of variance at the between-unit level of interest.}
}
\value{
The powRICLPM \code{object} with the \code{parameter} element in \code{object$session} set to the value of the \code{parameter} argument.
}
\description{
S3 method for class `powRICLPM`. \code{summary.powRICLPM} summarizes and outputs the setup and results of the powRICLPM analysis either generally, for a specific parameter (when \code{parameter} argument is specified), or for a specific parameter in a specific experimental condition (when the \code{parameter}, \code{sample_size}, \code{time_points}, and \code{ICC} arguments are specified). Optionally, it can display the parameter names contained in \code{object}.
}
\details{
\code{summary.powRICLPM} sets the \code{parameter} element in \code{session} with the value of the \code{parameter} argument, and returns the \code{object}. Saving this object might be useful as it renders the \code{parameter} argument for \code{\link{plot.powRICLPM}} and \code{\link{coef.powRICLPM}} superfluous.
\subsection{Parameter names}{When simulating the power across conditions with a varying number of time points, there are also a different numbers of parameters across the experimental conditions. By specifying the argument \code{names = TRUE}, this function returns the parameter names from the condition with the smallest number of parameters. As such, the returned parameter names are valid for each experimental condition.}
}
\examples{
# Example - Same starting point as example from ?powRICLPM()
# Define population parameters for lagged effects and within-component correlations
Phi <- matrix(c(.4, .1, .2, .3), ncol = 2, byrow = TRUE)
wSigma <- matrix(c(1, .3, .3, 1), ncol = 2, byrow = TRUE)

# Setup parallel computing (multicore, speeding up the analysis)
\dontrun{
library(furrr)
plan(multisession, workers = 4)

# Run the analysis
out1 <- powRICLPM(
  target_power = 0.8,
  search_lower = 500,
  search_upper = 1000,
  search_step = 100,
  time_points = c(3, 4),
  ICC = c(0.4, 0.5, 0.6),
  RI_cor = 0.3,
  Phi = Phi,
  wSigma = wSigma,
  reps = 100,
  seed = 123456
)

# General results (i.e., setup of powRICLPM analysis)
summary(out)

# General results and recommendation for specific parameter
summary(out, parameter = "wB2~wA1")

# Set parameter element of object
out_wB2wA1 <- summary(out, parameter = "wB2~wA1")
}
}
