% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.powRICLPM}
\alias{summary.powRICLPM}
\title{powRICLPM Power Analysis Summary}
\usage{
\method{summary}{powRICLPM}(object, ..., parameter = NULL)
}
\arguments{
\item{object}{A "powRICLPM" object.}

\item{...}{Phantom argument that has no influence on the summary produced. Included here to make the generic \code{summary()} a method for objects of class "powRICLPM".}

\item{parameter}{A character string denoting a single variable of interest. Use \code{\link{names.powRICLPM}} to get an overview of valid parameter names for the powRICLPM object.}
}
\description{
\code{summary.powRICLPM()} is a method for class "powRICLPM", summarizing the results of an à priori or post hoc powRICLPM simulation study.
}
\examples{
# Define population parameters for lagged effects and within-component correlations
Phi <- matrix(c(.4, .1, .2, .3), ncol = 2, byrow = T)
wSigma <- matrix(c(1, .3, .3, 1), ncol = 2, byrow = T)

# Create powRICLPM object for à priori power analysis
output <- powRICLPM(target_power = 0.5,
                     parameter = "wB2~wA1",
                     search_lower = 300,
                     search_upper = 600,
                     search_step = 50,
                     time_points = 3,
                     ICC = 0.5,
                     RI_cor = 0.3,
                     Phi = Phi,
                     wSigma = wSigma,
                     reps = 50,
                     cores = 1,
                     seed = 123456)

# Summarize results
summary(output, parameter = "wB2~wA1")
}
\seealso{
\code{\link{names.powRICLPM}} to get an overview of valid parameter names included in the powRICLPM object.
}
